dashboard/requirements.txt
Flask
paho-mqtt
cryptography